% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGisData.R
\name{addGisData}
\alias{addGisData}
\title{adds the Gis data to downloaded dataset}
\usage{
addGisData(x, host = "https://dw.euro.who.int")
}
\arguments{
\item{x}{a downloaded dataset created by \link{getDatasets}}

\item{host}{the WHO/EU host (use default)}
}
\value{
a GIS annotated dataset
}
\description{
adds the Gis data to downloaded dataset
}
\examples{
a=addGisData(getDatasets(c("HFA_13","HFA_426")))
plot(subset(a,code=="HFA_13")["value"])

table(a$iso3)
table(a$year)

data(world)
b=subset(a,code=="HFA_426" & year == 2011)

ggplot()+
  scale_fill_distiller(palette = "Spectral") +
  geom_sf(data=world,fill="white") + 
  geom_sf(data=b,aes(fill=value)) +
  xlim(c(-10,50))+ ylim(c(35,70)) +
  ggtitle(b[1,"full_name"],
          subtitle=b[1,"year"])

# for some reason have problems with the region clipping
data(world)
b=subset(a,code=="HFA_13" & year \%in\% c(2011,2015))
ggplot()+
  facet_grid(year~sex) +
  scale_fill_distiller(palette = "Spectral") +
  geom_sf(data=world,fill="white") + 
  geom_sf(data=b,aes(fill=value)) +
  xlim(c(-10,50))+ ylim(c(30,70)) +
  ggtitle(b[1,"full_name"],
          subtitle=b[1,"year"])
}
