% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eudata-package.R
\docType{package}
\name{eudata-package}
\alias{eudata}
\alias{eudata-package}
\title{eudata: Downloads data from gisco-services.ec.europa.eu}
\description{
gisco-services.ec.europa.eu provides data related to the european union through its rest API. This package tries to make it easier to get these data into R.
}
\author{
\strong{Maintainer}: Vilmos Prokaj \email{prokaj.vilmos@gmail.com}

}
\keyword{internal}
