% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets}
\alias{get_datasets}
\title{Retrieve available datasets from an endpoint}
\usage{
get_datasets(api)
}
\arguments{
\item{api}{An endpoint}
}
\value{
A tibble of available datasets.
}
\description{
This function returns the list of available datasets as a tibble.
The columns of the tibble provide information about each dataset.
}
\examples{
get_topic("Coastal lines") |>
  get_datasets()

}
