% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_files}
\alias{get_latest_files}
\title{Retrieve the latest files from the API}
\usage{
get_latest_files(api)
}
\arguments{
\item{api}{An endpoint to the dataset.}
}
\value{
A named list of files.
}
\description{
This function retrieves the files belonging to the
latest version of the given dataset.
When the dataset is not updated the cached version is returned.
}
\examples{
get_latest_files(get_topic("Postal"))$csv
}
