% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_topic}
\alias{get_topic}
\title{Retrieve Topic Information}
\usage{
get_topic(topic)
}
\arguments{
\item{topic}{A string representing the topic to retrieve.}
}
\value{
A request object to the specific endpoint.
}
\description{
This function fetches the details of a specific topic based
on the provided topic name.
}
\examples{
get_topic("Coastal lines")
get_topic("Postal")
}
