% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_topics}
\alias{get_topics}
\title{Retrieve Topics from API}
\usage{
get_topics()
}
\value{
A tibble the topics retrieved from the API.
}
\description{
Retrieves a list of topics from the specified API
endpoint `https://gisco-services.ec.europa.eu/distribution/v2/`.
}
\details{
This function sends a request to the given API endpoint
and parses the response to extract topic information.
}
\examples{
# Retrieve topics from the default endpoint
topics <- get_topics()
}
