% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleDataset.r
\name{getSingleDataset}
\alias{getSingleDataset}
\title{downloads a single dataset from the API}
\usage{
getSingleDataset(code = "HFA_13", host = "https://dw.euro.who.int")
}
\arguments{
\item{code}{the HFA code for the data set}

\item{host}{the WHO/EU host (use default)}
}
\value{
a dataset of values and dimensions
}
\description{
Downloads the data and dimensions (country, year etc.) for a single code.
Reports only single country data (returning an iso3 country code) 
and converts values and years into numerical types.
}
\examples{
a=getSingleDataset("HFA_13")
str(a)
a=getSingleDataset("HFA_426")
str(a)
}
