% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoPanel.r
\name{infoPanel}
\alias{infoPanel}
\title{displays basic information about the app}
\usage{
infoPanel(id, input = NULL, output = NULL, ui = T)
}
\arguments{
\item{id}{}

\item{input}{}

\item{output}{}

\item{ui}{}
}
\description{
displays basic information about the app
}
