% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleView.r
\name{simpleView}
\alias{simpleView}
\title{plots a simple panel for given codes and years}
\usage{
simpleView(id, input = NULL, output = NULL, ui = T, cx = NULL)
}
\arguments{
\item{id}{the widget id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{ui}{TRUE for the user interface, FALSE for server}

\item{cx}{a reactive list, reads data from cx$egData}
}
\description{
plots a simple panel for given codes and years
}
