% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderView.r
\name{sliderView}
\alias{sliderView}
\title{plots a simple slider view for given codes and years}
\usage{
sliderView(id, input = NULL, output = NULL, ui = T, cx = NULL)
}
\arguments{
\item{id}{the widget id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{ui}{TRUE for the user interface, FALSE for server}

\item{cx}{a reactive list, reads data from cx$egData}
}
\description{
Provides two controls - one to select the endpoint and the other 
to select the year
}
