% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eudata-package.R
\docType{package}
\name{eudata-package}
\alias{eudata}
\alias{eudata-package}
\title{eudata: Downloads Data from Gisco-services.ec.europa.eu}
\description{
The service gisco-services.ec.europa.eu provides data related to the european union through its rest API. This package tries to make it easier to get these data into R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prokaj/eudata}
  \item Report bugs at \url{https://github.com/prokaj/eudata/issues}
}

}
\author{
\strong{Maintainer}: Vilmos Prokaj \email{prokaj.vilmos@gmail.com}

}
\keyword{internal}
