% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eudata-package.R
\docType{package}
\name{eudata-package}
\alias{eudata}
\alias{eudata-package}
\title{eudata: Access Data from 'GISCO'}
\description{
Access data related to the European union from 'GISCO' \url{https://ec.europa.eu/eurostat/web/gisco}, the Geographic Information System of the European Commission, via its rest API at \url{https://gisco-services.ec.europa.eu}. This package tries to make it easier to get these data into R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prokaj/eudata}
  \item Report bugs at \url{https://github.com/prokaj/eudata/issues}
}

}
\author{
\strong{Maintainer}: Vilmos Prokaj \email{prokaj.vilmos@gmail.com}

}
\keyword{internal}
