% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.r
\name{getDatasets}
\alias{getDatasets}
\title{download multiple datasets by code}
\usage{
getDatasets(codes = c("HFA_13", "HFA_426"),
  host = "https://dw.euro.who.int")
}
\arguments{
\item{codes}{a vector of HFA codes}

\item{host}{the WHO/EU host (use default)}
}
\value{
a dataset of values and dimensions
}
\description{
You can get indicator codes by browsing from
\url{https://gateway.euro.who.int/en/datasets/european-health-for-all-database/}
and looking at the head of the .csv file.
}
\examples{
a=getDatasets(c("HFA_13","HFA_426"))
str(a)
with(a,table(name,sex))
}
